move pagopacapability
